sig fix

+++ src/RandomisationMan.h
@@ -19,7 +19,7 @@ private:
 	//This function template is called by external game code
 	//Don't touch the signature of this function. 
 	template<std::unique_ptr<Randomizer>* rnd>
-	static int __fastcall pushWorldItem(__int64* worldInventory, const RepositoryID* repoID, __int64 a3, void* a4, __int64 a5, __int64 a6, __int64* a7, void* a8, char* a9, char a10) {
+	static __int64 __fastcall pushWorldItem(__int64* worldInventory, const RepositoryID* repoID, __int64 a3, void* a4, __int64 a5, __int64 a6, __int64* a7, void* a8, char* a9, char a10) {
 		const RepositoryID* id = (*rnd)->randomize(repoID);
 		return reinterpret_cast<tPushItem>(GameOffsets::instance()->getPushItem())(worldInventory, id, a3, a4, a5, a6, a7, a8, a9, a10);
 	};
